Crate storage_service_client

Structs

StorageServiceClient

The interface for sending Storage Service requests and querying network peer information.

StorageServiceNetworkSender

The Storage Service network sender for a single network.

Enums

Error

Functions

network_endpoint_config

Type Definitions

StorageServiceMultiSender

A network sender that dispatches across multiple networks.